more stuff

+++ lib/Ui/Screens/Orders/order_widget.dart
@@ -3,7 +3,8 @@ import 'package:online_order_shop_mobile/Application/Providers/navigation_provid
 import 'package:online_order_shop_mobile/Domain/Orders/iorder.dart';
 import 'package:online_order_shop_mobile/Domain/Orders/order_status.dart';
 import 'package:online_order_shop_mobile/Infrastructure/service_provider.dart';
-import 'package:online_order_shop_mobile/Ui/Components/dialogs.dart';
+import 'package:online_order_shop_mobile/Ui/Components/Dialogs/confirmation_dialog.dart';
+import 'package:online_order_shop_mobile/Ui/Components/Dialogs/dialogs.dart';
 import 'package:online_order_shop_mobile/Ui/Themes/constants.dart';
 import 'package:provider/provider.dart';
 
@@ -57,9 +58,18 @@ class _OrderWidgetState extends State<OrderWidget> {
                     color: theme.colorScheme.error, fontSize: textSizeMeduim2),
               ),
               onPressed: () {
-                ServicesProvider()
-                    .orderService
-                    .deleteOrder(widget.order.getId());
+                showDialog<AlertDialog>(
+                    context: context,
+                    builder: (context) {
+                      return ConfirmAlertDialog(
+                        onConfirm: () {
+                          ServicesProvider()
+                              .orderService
+                              .deleteOrder(widget.order.getId());
+                        },
+                        message: messagePermanantAction,
+                      );
+                    });
               },
             ),
           ),
